Use `unsafe` only where necessary

+++ vm/src/vm/mod.rs
@@ -768,20 +768,15 @@ impl VirtualMachine {
                 self.print_exception(exc);
                 #[cfg(unix)]
                 {
-                    unsafe {
-                        if sigaction(
-                            SIGINT,
-                            &SigAction::new(
-                                nix::sys::signal::SigHandler::SigDfl,
-                                SaFlags::SA_ONSTACK,
-                                SigSet::empty(),
-                            ),
-                        )
-                        .is_ok()
-                        {
-                            interpreter::flush_std(self);
-                            kill(getpid(), SIGINT).expect("Expect to be killed.");
-                        }
+                    let action = SigAction::new(
+                        nix::sys::signal::SigHandler::SigDfl,
+                        SaFlags::SA_ONSTACK,
+                        SigSet::empty(),
+                    );
+                    let result = unsafe { sigaction(SIGINT, &action) };
+                    if result.is_ok() {
+                        interpreter::flush_std(self);
+                        kill(getpid(), SIGINT).expect("Expect to be killed.");
                     }
 
                     (libc::SIGINT as u8) + 128u8
